update auth export and import to include emailtemplatestore

+++ auth/export.sh
@@ -14,11 +14,19 @@ docker cp "$HABIDAT_DOCKER_PREFIX-user":/habidat-user/data/emailTemplateStore.js
 sed -f export.sed $HABIDAT_BACKUP_DIR/$HABIDAT_DOCKER_PREFIX/auth/export.ldif.tmp > $HABIDAT_BACKUP_DIR/$HABIDAT_DOCKER_PREFIX/auth/export.ldif
 rm $HABIDAT_BACKUP_DIR/$HABIDAT_DOCKER_PREFIX/auth/export.ldif.tmp
 echo "Compressing data..."
-tar -czf auth-$DATE.tar.gz -C $HABIDAT_BACKUP_DIR/$HABIDAT_DOCKER_PREFIX/auth export.ldif -C $HABIDAT_BACKUP_DIR/$HABIDAT_DOCKER_PREFIX/auth activationStore.json -C $HABIDAT_BACKUP_DIR/$HABIDAT_DOCKER_PREFIX/auth emailTemplateStore.json
+if [ -f $HABIDAT_BACKUP_DIR/$HABIDAT_DOCKER_PREFIX/auth/emailTemplateStore.json ]
+then
+  tar -czf auth-$DATE.tar.gz -C $HABIDAT_BACKUP_DIR/$HABIDAT_DOCKER_PREFIX/auth export.ldif -C $HABIDAT_BACKUP_DIR/$HABIDAT_DOCKER_PREFIX/auth activationStore.json -C $HABIDAT_BACKUP_DIR/$HABIDAT_DOCKER_PREFIX/auth emailTemplateStore.json
+else 
+  tar -czf auth-$DATE.tar.gz -C $HABIDAT_BACKUP_DIR/$HABIDAT_DOCKER_PREFIX/auth export.ldif -C $HABIDAT_BACKUP_DIR/$HABIDAT_DOCKER_PREFIX/auth activationStore.json
+fi
 mv auth-$DATE.tar.gz $HABIDAT_BACKUP_DIR/$HABIDAT_DOCKER_PREFIX/auth/
 rm $HABIDAT_BACKUP_DIR/$HABIDAT_DOCKER_PREFIX/auth/export.ldif
 rm $HABIDAT_BACKUP_DIR/$HABIDAT_DOCKER_PREFIX/auth/activationStore.json
-rm $HABIDAT_BACKUP_DIR/$HABIDAT_DOCKER_PREFIX/auth/emailTemplateStore.json
+if [ -f $HABIDAT_BACKUP_DIR/$HABIDAT_DOCKER_PREFIX/auth/emailTemplateStore.json ]
+then
+  rm $HABIDAT_BACKUP_DIR/$HABIDAT_DOCKER_PREFIX/auth/emailTemplateStore.json
+fi
 
 echo "NOTE: importing this data only works for data with the same domain / LDAP base"
